refactor: harden ECS migration boundaries
Amp-Thread-ID: https://ampcode.com/threads/T-01a034c8-4c01-75a8-a372-eede89c8fc37
